Arrange Builder Factory test files
